Restructure task (#9)
* File and Folder Restructure

Moved files and folder structure to conform with conventional pattern.

* Update Readme

Minor edits that reflects changes in the project.

* Remove Double Quotes

Minor fix.  Removed double quotes on a include line.

* Create Package

Create a config.cmake file for this library.

* d postfix

Add a d postfix to the library name so the client code can link to the correct binary.

* Update Readme

Added Build and Install section.

* Update Version Number

Update version number to 1.0.0.